Move the hot-reload suite out of the gallery entry it was living in
`docs/demo-programs.md` defines a gallery entry as a `program.cambra` plus a `mod.rs` with one or more `#[test]` functions, and `tests/programs/hot_reload/mod.rs` was 3,309 lines of feature suite — 58 cases over roughly twenty guard-table rows across the induction, transactional and `stdin` domains, larger than any source file the feature has. The cases drive `LiveProgram::start`, `drive_main_to_terminal` and `launch_under_control` rather than compiling and evaluating, so they are not `compilation_pipeline` tests either. The slot the layout already has is a top-level suite: `tests/hot_reload.rs` beside `tests/compilation_pipeline.rs` and `tests/http_server.rs`, with `tests/hot_reload/harness.rs` holding the programs and the plumbing and `tests/hot_reload/cases.rs` the cases.

The gallery entry keeps the two `.cambra` files and the one case they are for, which is the demonstration: sign the guestbook twice, edit the accumulating loop, and see the third signature answered under the new rule with the first two still standing. That case is now the entry's whole `mod.rs`, and the entry has the row in `docs/demo-programs.md` its own policy asks for.

Two binaries now run a program and talk to it, so what both need moves to `tests/support/serving.rs`: the sink compile, the scheduler pump, and the `TcpStream` HTTP client. `tests/programs/common/mod.rs` keeps the gallery's expectation helpers, which nothing else wants. The split is what makes each binary use all of what it imports — `common` compiled into the hot-reload binary would report its eight gallery-only helpers as dead.

`exchange` was in the hot-reload module and is the collapsed form of a shape three other entries hand-roll: a channel, a client thread, and `drive_until` on the response. It moves to `serving` and those three use it, which is also what makes it worth having rather than a fourth copy.

`storefront/mod.rs` and `demo-programs.md` described the version dimension as open design awaiting a `Versioned` surface. Replacing one version with another is implemented, so what is open is serving two at once, and both say that instead.
